feat: add command get match day
