+++ src/commands/index.ts
@@ -4,6 +4,7 @@ export * from "./banner";
 export * from "./cat";
 export * from "./dice";
 export * from "./dog";
+export * from "./getMatchDay";
 export * from "./icon";
 export * from "./kick";
 export * from "./love";
